+++ src/api/SFA.DAS.Provider.Events.Application/DataLock/GetDataLockEventsQuery/GetDataLockEventsQueryRequest.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.Provider.Events.Application.DataLock.GetDataLockEventsQuery
 {
     public class GetDataLockEventsQueryRequest : IAsyncRequest<GetDataLockEventsQueryResponse>
     {
-        public int SinceEventId { get; set; }
+        public long SinceEventId { get; set; }
         public DateTime? SinceTime { get; set; }
         public string EmployerAccountId { get; set; }
         public long Ukprn { get; set; }
